Modelling RCU with fences instead of special events
