+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/arch/linux/event/RMWFetchOp.java
@@ -1,4 +1,4 @@
-package com.dat3m.dartagnan.program.arch.linux.event.rmw;
+package com.dat3m.dartagnan.program.arch.linux.event;
 
 import com.dat3m.dartagnan.expression.ExprInterface;
 import com.dat3m.dartagnan.expression.IExpr;
@@ -18,31 +18,31 @@ import com.dat3m.dartagnan.wmm.utils.Arch;
 import java.util.Arrays;
 import java.util.LinkedList;
 
-public class RMWOpReturn extends RMWAbstract implements RegWriter, RegReaderData {
+public class RMWFetchOp extends RMWAbstract implements RegWriter, RegReaderData {
 
     private final IOpBin op;
 
-    public RMWOpReturn(IExpr address, Register register, ExprInterface value, IOpBin op, String mo) {
+    public RMWFetchOp(IExpr address, Register register, ExprInterface value, IOpBin op, String mo) {
         super(address, register, value, mo);
         this.op = op;
     }
 
-    private RMWOpReturn(RMWOpReturn other){
+    private RMWFetchOp(RMWFetchOp other){
         super(other);
         this.op = other.op;
     }
 
     @Override
     public String toString() {
-        return resultRegister + " := atomic_" + op.toLinuxName() + "_return" + Mo.toText(mo) + "(" + value + ", " + address + ")";
+        return resultRegister + " := atomic_fetch_" + op.toLinuxName() + Mo.toText(mo) + "(" + value + ", " + address + ")";
     }
 
     // Unrolling
     // -----------------------------------------------------------------------------------------------------------------
 
     @Override
-    public RMWOpReturn getCopy(){
-        return new RMWOpReturn(this);
+    public RMWFetchOp getCopy(){
+        return new RMWFetchOp(this);
     }
 
 
@@ -52,12 +52,18 @@ public class RMWOpReturn extends RMWAbstract implements RegWriter, RegReaderData
     @Override
     public int compile(Arch target, int nextId, Event predecessor) {
         if(target == Arch.NONE) {
-            Register dummy = new Register(null, resultRegister.getThreadId());
-            RMWLoad load = new RMWLoad(dummy, address, Mo.loadMO(mo));
-            Local local = new Local(resultRegister, new IExprBin(dummy, op, value));
-            RMWStore store = new RMWStore(load, address, resultRegister, Mo.storeMO(mo));
+            Register dummy = resultRegister;
+            if(resultRegister == value){
+                dummy = new Register(null, resultRegister.getThreadId());
+            }
 
-            LinkedList<Event> events = new LinkedList<>(Arrays.asList(load, local, store));
+            RMWLoad load = new RMWLoad(dummy, address, Mo.loadMO(mo));
+            RMWStore store = new RMWStore(load, address, new IExprBin(dummy, op, value), Mo.storeMO(mo));
+
+            LinkedList<Event> events = new LinkedList<>(Arrays.asList(load, store));
+            if (dummy != resultRegister) {
+                events.addLast(new Local(resultRegister, dummy));
+            }
             if (mo.equals(Mo.MB)) {
                 events.addFirst(new Fence("Mb"));
                 events.addLast(new Fence("Mb"));
